cm-t335: Introde CM-T335 temporary support code
CM-T335 support code is temporary as it is used only for package 4.5.
Most of the common files are deleted by an empty version.
Program script for NAND is replacing the common (for eMMC) program script.
The support code is implemented in the CM-T335 directory and not by
system design because of tight schedule.
Next version should be implemented in common code.

Signed-off-by: Uri Mashiach <uri.mashiach@compulab.co.il>