cm-t43: bootscr: update first run parameters
Update first run parameters to make production easy (no need for
production worker to remove the SD-Card befor rebooting the system),
and also adjust the default boot cmd to boot from emmc as if there
was no SD-Card inserted. This adjustment is for the new installation
method (on emmc rather than SD-Card) that assumes that the customer
will just plug in the eval kit without inserting any SD-Card into it.

Signed-off-by: Nikita Kiryanov <nikita@compulab.co.il>