NAND_PARAMS: fix early return
The variable NAND_PARAMS contains the NAND program parameters.
In case the NAND is not used the variable should be empty.

The following functions return early if the NAND_PARAMS variable is
empty: unmount_destination, format_partitions, copy_kernel_files.

The correct behaviour is to skip the NAND function execution.

Signed-off-by: Uri Mashiach <uri.mashiach@compulab.co.il>