imx7: bootscr: fix final bootcmd content
The boot script is setting the value of the next bootcmd and the final
value.

Currently the value of the environment variable bootcmd_next is set to
empty. Therefor the final value of the environment variable bootcmd is
empty.
Automatic boot is prevented, starting from the second boot after the
installation.

Fix the problem by adding double quotes to the update command of the
environment variables bootcmd_next and bootcmd.

Signed-off-by: Uri Mashiach <uri.mashiach@compulab.co.il>