bootscr: add SD mmc device detection
In most of the platforms the mmc device number of the SD card is stored
in a variable before running the boot script.

Add an mmc device detection mechanism, in case that the mmc device is
not stored in the environment.
The mmc device is set to the number of the first mmc device which
includes the platform configuration file.

Signed-off-by: Uri Mashiach <uri.mashiach@compulab.co.il>