common: remove the platform directory support
The platform directory includes boot script and installation scripts
which are relevant for specific platforms.

All of the platform parameters were transferred to the platform
configuration files.

Remove the platform directory support from the tools scripts.
Includes the binary output of the tools which are common to all of the
platforms.

Signed-off-by: Uri Mashiach <uri.mashiach@compulab.co.il>